changed to file per tenant
Also set it up to add the allowed login types and to make MFA a enabled/disabled/unknown just like the other columns we just added.
Also cleaned up the code to be more DRY.
Also added a way to only run against a  single tenant for testing or quick updates.